fix: fast-path worker-health to skip CLI bootstrap (#1974)
`vibetuner worker-health` paid the full CLI/app bootstrap (sub-command
imports plus load_app_config importing the user's tune.py) before it
ever read the streaq health key. On a real deploy that bootstrap takes
~5s, so every probe was killed at the scaffolded healthcheck's 5s
timeout and a healthy worker flapped to permanently unhealthy.

Move the console entry to a main() that dispatches `worker-health`
straight to a lightweight check (config + Redis client only) before the
heavy Typer app is imported. The full app, with every command and the
user CLI, now lives in vibetuner.cli.root and is still built lazily for
all other invocations. The healthcheck returns in well under a second.
