+++ vibetuner-py/src/vibetuner/__main__.py
@@ -1,4 +1,6 @@
-from .cli import app
+# ABOUTME: Enables `python -m vibetuner`, sharing the console script entry point.
+# ABOUTME: Routes through main() so the worker-health fast path applies here too.
+from .cli import main
 
 
-app()
+main()
