+++ vibetuner-template/pyproject.toml.j2
@@ -20,7 +20,7 @@ dependencies = [
 ]
 
 [project.scripts]
-vibetuner = "vibetuner.cli:app"
+vibetuner = "vibetuner.cli:main"
 
 [dependency-groups]
 dev = [
